feat: Define Post schema in MongoDB
This commit introduces a new schema for posts in MongoDB using Mongoose. The `postSchema` is defined with the following fields:

- userId (required)
- firstName (required)
- lastName (required)
- location
- description
- picturePath
- userPicturePath
- likes (a map of Boolean)
- comments (default to an empty array)

In addition, the schema is configured with timestamps enabled, automatically adding 'createdAt' and 'updatedAt' fields to each document.

A Mongoose model 'Post' is created from this schema and exported for use in other parts of the application.

File(s) updated:
- server/models/Post.js