feat: Add dummy pictures to public assets
This commit includes the addition of dummy images to the server/public/assets directory. These images will be used for displaying as profile pictures and post pictures in the dummy data for users and posts respectively. This addition is crucial for visually testing the UI and to ensure correct image path resolution and display.

File(s) added:
- public/assets/p1.jpeg
- public/assets/p2.jpeg
- public/assets/p3.jpeg
- public/assets/p4.jpeg
- public/assets/p5.jpeg
- public/assets/p6.jpeg
- public/assets/p7.jpeg
- public/assets/p8.jpeg
- public/assets/p9.jpeg
- public/assets/p10.jpeg
- public/assets/p11.jpeg
- public/assets/p12.jpeg
- public/assets/p13.jpeg
- public/assets/post1.jpeg
- public/assets/post2.jpeg
- public/assets/post3.jpeg
- public/assets/post4.jpeg
- public/assets/post5.jpeg
- public/assets/post6.jpeg
- public/assets/twitter.png
- public/assets/linkedin.png
- public/assets/info1.jpeg
- public/assets/info2.jpeg
- public/assets/info3.jpeg
- public/assets/info4.jpeg