feat: Add new client folder with front-end components
This commit includes the addition of a new client folder that contains all the necessary front-end components of the app. The addition of this folder marks the beginning of front-end development for the app and serves as the base for the user interface.

Directory added:
- client/

This client directory contains various subdirectories and files crucial for the front-end development, including React components, stylesheets, utility functions, and configurations.